First attempt at automating packet loss graph generation.

+++ FPGA/PacketDropperTopSDx/Build_project.bash
@@ -1,4 +1,10 @@
 #!/bin/bash -e
+if [ -z "$DISPLAY" ]
+then
+  echo "Despite running SDx in batch mode, you have to set the DISPLAY environment variable."
+  exit 1
+fi
+
 rm -fr $PWD/PacketDropperTopSDx
 mkdir -p Sources
 sed "s/RATE/0/" Dropper_init_template.c > Sources/Dropper_init.c
